+++ CHANGELOG.md
@@ -10,6 +10,8 @@ the [discussions section](https://github.com/foresterre/cargo-msrv/discussions).
 
 ## Unreleased
 
+## 0.19.0 - 2026-03-01
+
 ### Added
 
 * Added `--skip-unavailable-toolchains` which treats a Rust version as incompatible when a toolchain failed to install or was otherwise unavailable
@@ -25,7 +27,7 @@ the [discussions section](https://github.com/foresterre/cargo-msrv/discussions).
 ### Changed
 
 * Improve compatibility of path canonicalization on Windows
-* Restore lockfile moved by `--ignore-lockfile`, after a user interrupts the process
+* Restore lockfile when moved by `--ignore-lockfile`, after a user interrupts the process
 
 ## 0.18.3 - 2025-03-12